+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `FieldModifiers#modify(Function<String, String>)` to simply modify fields via functional interface
+
+### Deprecated
+- `SimpleFieldModifier` interface
+- `FieldModifiers#lower` and `FieldModifiers#upper` methods
+
 ## [3.6.0] - 2025-03-04
 ### Added
 - Configuration of maximum fields, maximum field size, and maximum record size via record handler
@@ -154,6 +162,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
+[Unreleased]: https://github.com/osiegmar/FastCSV/compare/v3.6.0...HEAD
 [3.6.0]: https://github.com/osiegmar/FastCSV/compare/v3.5.0...v3.6.0
 [3.5.0]: https://github.com/osiegmar/FastCSV/compare/v3.4.0...v3.5.0
 [3.4.0]: https://github.com/osiegmar/FastCSV/compare/v3.3.1...v3.4.0